pyproject.toml
[project]
name = "aiworks"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web / API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    # DB
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    # Settings
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    # Agentic stack
    "langchain>=0.2",
    "langchain-openai>=0.1",
    "langgraph>=0.1",
    # Scraping — crawl4ai (Playwright, handles CSR/SSR/SPA) + sync fallback
    "crawl4ai>=0.4",
    "playwright>=1.44",
    "requests>=2.31",
    "trafilatura>=1.9",
    # Observability
    "opentelemetry-sdk>=1.24",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "langfuse>=3.0",
    "deepeval>=3.9.6",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.black]
target-version = ["py311"]
line-length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app", "agents", "tools"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "llm: tests that call external LLM APIs (set OPENAI_API_KEY)",
]
